qemu: add delay time cpu stats
This commit adds delay time (steal time inside guest) to libvirt
domain CPU stats. It's more convenient to work with this statistic
in a context of libvirt domain. Any monitoring software may use
this information.

As an example: the next step is to add support to
libvirt-go and expose metrics with libvirt-exporter.

Signed-off-by: Aleksei Zakharov <zaharov@selectel.ru>
Message-Id: <20210129143402.630-1-zaharov@selectel.ru>