log error if virConnectCacheOnceInit() fails
This may be pointless, but I noticed it and it was bugging me.

virGetConnectNetwork() calls
 virGetConnectGeneric(), which calls
  virConnecCacheInitialize(), which is actually a call (only once) to
   virConnectCacheOnceInit() which calls
    virThreadLocalInit() several times, which calls
     pthread_key_create()

If pthread_key_create() fails, it (of course) doesn't log an error
(because it's not a part of libvirt), nor does any other function on
the call chain all the way up to virGetConnectNetwork(). But none of
the callers of virGetConnectNetwork() log an error either, so it is
possible that an API could fail due to virGetConnectNetwork() failing,
but would only log "an error was encountered, but the cause is
unknown. Deal with it."  (paraphrasing).

In all likelyhood, virConnectCacheOnceInit() is going to be called at
some earlier time, and almost certainly pthread_key_create() will
never fail (and if it does, the user will have *much* bigger problems
than an obtuse error message from libvirt). So I don't know if there's
any value at all in pushing this. Just throwing it out there...

Signed-off-by: Laine Stump <laine@redhat.com>
Message-Id: <20210201062720.425041-2-laine@redhat.com>
Reviewed-by: Michal Privoznik <mprivozn@redhat.com>