storage_backend_iscsi(_direct): Properly clear secrets
The code pretends that it cares about clearing the secret values, but
passes the secret value to a realloc, which may copy the value somewhere
else and doesn't sanitize the original location when it does so.

Since we want to construct a string from the value, let's copy it to a
new piece of memory which has the space for the 'NUL' byte ourselves, to
prevent a random realloc keeping the data around.

While at it, use memset of VIR_DISPOSE_N since it's being phased out.

Signed-off-by: Peter Krempa <pkrempa@redhat.com>
Message-Id: <76e0a05bb83fb2b98e4cb5890d1ac85ee7dcb374.1612186636.git.pkrempa@redhat.com>