ci: Expose CI_USER_LOGIN as a Makefile variable for users to use
More often than not I find myself debugging the containers which means
that I need to have root inside, but without manually tweaking the
Makefile each time the execution would simply fail thanks to the
uid/gid mapping we do. What if we expose the CI_USER_LOGIN variable,
so that when needed, the root can be simply passed with this variable
and voila - you have a root shell inside the container with CWD=~root.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <5a9da8e2091c887c58a50ec7a592e3fdfdcb293e.1612976435.git.eskultet@redhat.com>