ci: Makefile: Specify a help target to replace ci-help
It's quite pointless to have a 'ci-help' target in the Makefile when
one needs to actually open the Makefile to go through the list of
targets to know what functionality the Makefile actually provides. It's
much more intuitive to run "make help" in that case. Therefore, add a
'help' target and replace the old 'ci-help' target with it.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <c8a3a90eca20356d5679a011ce64f7f20a7108e8.1612977855.git.eskultet@redhat.com>