ci: Introduce a new checker script 'check-registry.py'
When removing the dependency on sudo in our CI Makefile, I realized
that the list of images that we pull contains some old images that we
no longer support in lcitool, but can be still pulled from the
registry. This is a simple checker which runs in context of the refresh
script which informs the developer/maintainer that there are some old
redundant images that should be purged from the registry, so that
people don't try to run local builds in those (which will most likely
fail anyway).

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <765ef45d8b45eb6adb54beed9e92273d651b7f9c.1612977855.git.eskultet@redhat.com>