clang-tidy: Allow timeouts
With an empty cache, the first run of the clang-tidy scan in the CI will
fail. While the instinctive reaction "press the rerun button" will eventually
lead to the situation where enough files are cached that the entire scan fits
within the time window, this creates friction for e.g. new contributors.

By allowing timeouts to be non-fatal events, we reduce scanning to a "best
effort" approach, that might miss newly introduced regressions.

Signed-off-by: Tim Wiederhake <twiederh@redhat.com>
Message-Id: <20210212132534.29066-7-twiederh@redhat.com>