clang-tidy: Add shuffle
Randomizing the order of files to scan has no impact for local use of the
script. The same holds true for use in the CI, if the amount of cached
files is big enough for the entire scan to finish before timeout.

If the cache is empty or not filled enough to ensure timely completion,
randomizing the order of files makes it more likely to spent time on caching
new files rather than hashing already cached files to check for the presence
of a cache file.

Signed-off-by: Tim Wiederhake <twiederh@redhat.com>
Message-Id: <20210212132534.29066-8-twiederh@redhat.com>